build: integrate Tailwind CSS for project styling
Add Tailwind CSS for styling and Lightning CSS for transforming and minifying CSS. Because of the way Electron handles Vite configs (and there being 4 of them) I wasn't sure which Vite config to add the Lightning CSS options to. I ended up adding them to the base config because adding them to renderer broke the styling. I'm not sure if Lightning CSS is actually transforming and minifying the CSS though so I'll have to figure out a way to do that later.