feat: filter out requests that do not match the types we want
Crawlee was enqueueing all of the links it found, including things like pdfs and images. We may want to do something with those later but for now they would just retry a few times and error because they were not added to additionalMimeTypes. We are now filtering the links that are given to Crawlee and removing non-html file types. This saves bandwidth and system resources.