fix: patch header-generator to pass the correct paths to BayesianNetwork
Information about the bug and how it was fixed - https://app.clickup.com/9009037304/v/dc/8cfnzzr-497/8cfnzzr-597

TL;DR - header-generator was using __dirname in the file paths that it was giving to the generative-bayesian-network package which was passing them to adm-zip. __dirname was resolving to the Electron asar file but we need it to resolve to the unpacked files.

Fixes: CRAWLER-1